-added changes to db readme. - added db_query.js to Backend for db query purposes - to built upon later
