+++ BackEnd/package.json
@@ -15,6 +15,7 @@
     "express": "^4.21.1"
   },
   "devDependencies": {
+    "mysql2": "^3.11.4",
     "nodemon": "^3.1.7"
   }
 }
