--- DataBase/.gitignore
@@ -1,40 +1,0 @@
-# dependencies
-/node_modules
-
-# build output
-/build
-
-# local env files
-.env
-.env.local
-.env.development.local
-.env.test.local
-.env.production.local
-
-# log files
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-
-# editor and OS files
-.DS_Store
-Thumbs.db
-*.sublime-project
-*.sublime-workspace
-*.code-workspace
-.vscode/
-.idea/
-
-# coverage directory used by testing tools
-/coverage
-
-# temporary files
-*.tmp
-*.temp
-
-# optional npm cache directory
-.npm
-
-# additional directories
-/.next
-/.expo
