--- DataBase/package.json
@@ -1,6 +1,0 @@
-{
-  "dependencies": {
-    "mysql": "^2.18.1",
-    "mysql2": "^3.11.4"
-  }
-}
